Fix reload plugin admin list check
The admins were being stored in an array but when the array was accessed
it was assumed the array contained the entire config structures and not just
the list of admins. Here we change the variable name to make it more clear
we are directly accessing a list of admins.